backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.13.3
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
email-validator==2.2.0
python-dotenv==1.0.1
Pillow==11.0.0
numpy==2.1.3
scipy==1.14.0
pydub==0.25.1
librosa==0.11.0
soundfile==0.13.1
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
slowapi==0.1.9
aiofiles==24.1.0
google-genai>=0.8.0
